pyproject.toml
[tool.poetry]
name = "jinja2-lucide"
version = "0.1.0"
description = "An jinja2 tags, allows lucide-icons to be embedded as inline svg into templates. "
authors = ["DotLabs Team <teams@dotlabs.dev>"]
packages = [{include = "jinja2_lucide.py"}]

[tool.poetry.dependencies]
python = "^3.14"
jinja2 = "^3.1.6"
lucide-py = {url = "https://github.com/dotlabshq/lucide-py/releases/download/0.2.1/lucide_py-0.2.1-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
